fixes complie issues
